+++ api/src/catalog/catalog.module.ts
@@ -1,9 +1,13 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Product } from './product.entity';
+import { CatalogController } from './catalog.controller';
+import { CatalogService } from './catalog.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Product])],
-  exports: [TypeOrmModule],
+  controllers: [CatalogController],
+  providers: [CatalogService],
+  exports: [TypeOrmModule, CatalogService],
 })
 export class CatalogModule {}