+++ api/src/clients/clients.module.ts
@@ -1,9 +1,13 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Client } from './client.entity';
+import { ClientsController } from './clients.controller';
+import { ClientsService } from './clients.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Client])],
-  exports: [TypeOrmModule],
+  controllers: [ClientsController],
+  providers: [ClientsService],
+  exports: [TypeOrmModule, ClientsService],
 })
 export class ClientsModule {}